Better seeking for IsoFile; using s64 to allow for a full 4gb of addressable file positions in either direction.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@2166 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/CDVD/IsoFS/IsoFile.h
@@ -26,7 +26,7 @@ public:
 	virtual ~IsoFile() throw();
 
 	u32 seek(u32 absoffset);
-	u32 seek(s32 offset, wxSeekMode ref_position);
+	u32 seek(s64 offset, wxSeekMode ref_position);
 	void reset();
 
 	s32 skip(s32 n);
